re1.5: Fix handling of escapes within character classes
In CircuitPython, we've added support for escaped characters in
character classes, like `r'[\r\n\t]'`.  On the micropython side
they added a simpler bit of code to skip over escapes within
character classes so that `[a\\b]' worked.  However, this very test
case was handled incorrectly by our own added code.

Basically, when a character class contains a *non-range*, the same
character of the pattern is parsed twice, to give the lower and upper
bounds of the range.  However, the code we added caused the regular
expression location to be advanced when the first backlash was encountered,
making r'[a\\b]' be parsed more like '[a\\-\b]' which was nonsense.

Track where parsing one character-or-range started (the new local variable
"b") and if a non-range is encountered, set the parsing position back to
where we started, which moves back over the backslash if necessary.

Also, add a check (in two places) for the degenerate regular expression
ending with a backslash inside of a character class, such as '[\\'
This fixes the bug, while preserving our new behavior.