runtime0: Renumber scope flags again
FLAG_ASYNC must be within FLAG_ALL_SIG, so it _does_ need to be
renumbered down to the 0x10 position, moving the other flags up
one bit.

FLAG_VIPERRET_POS must be shifted up 1, since (1<<7) is the next bit
after FLAG_HASCONSTS at 0x40.  This has to be done once in C and once
in Python.

This affects the encoding of scope_flags, causing them to spread over
an additional byte in the flexible-length integer encoding.  Everything
except the hard-coded mpy data in mpy_native.py accounts for this
automatically.

The fancy prefix-encoding of the function signature handles this,
though as each extension _byte_ only carries a single _flag bit_,
it takes 5 more bytes to specify an async function than a bog-standard
one.  Oh well.