Auto-configure display on Feather RP2350
This should be considered an RFC. Some bits should probably be
moved to common code & tidied up.

If connected to a HDMI monitor, an eeprom will be on the i2c bus
with address 0x50.

When it's found, configure the display to 320x240x16 (default), or to
user settings. Invalid user settings cause a safe mode with a misleading
error: "Heap allocation when VM not running." This is because
common_hal_picodvi_framebuffer_construct throws an exception for invalid
parameters (width, height, and color depth), but the circuitpy VM is not
running.

With the default framebuffer configuration, about 256KiB is left for the
CircuitPython VM.

When a display is configured (whether picodvi or not), board.DISPLAY
exits and can be retrieved. When there's no display (including after
`release_displays()`), accessing `board.DISPLAY` raises an AttributeError.
This behavior should maybe adapted everywhere, but I'm not sure how
much flash size it takes.