Fix starved audio on RP2
Switch to one control DMA channel and one data channel instead of
using background tasks to reset the read address of the DMA. When
the background task is starved, the DMA reads into other memory
because readaddr isn't reset. So, use a control channel to toggle
between the buffers by resetting the read address.