Fix DVI blanking during SPI transactions
SPI was using DMA to transfer to PSRAM. When a cache miss occurs,
the DMA can't switch to another transfer and throws off DVI timing.

So, only use DMA with SPI when the buffers are in SRAM. This will
slow down SPI transactions when the FIFOs are empty and the CPU is
busy running a background task. It will still be correct though
since we control the SPI clock.

Fixes #10557