error checks for usb device string properties
Improve the error handling for usb.core.Device string properties:
.serial_number, .product, .manufacturer

Previously, the property getters didn't check the device descriptor
to see if the device actually had the requested string. Instead,
they relied on TinyUSB to return a failure result if the string was
not available. That made it impossible to distinguish missing
strings from other more serious USB errors (e.g. unplugged device).

These changes make it possible to return None for a missing string
or raise a USBError exception in case of a more serious problem.